Rename api folder to extensions
