+++ src/testFile.ts
@@ -1,7 +1,7 @@
 import { commands, DocumentSymbol, LogLevel, SymbolKind, TestItem, TestRun, workspace, WorkspaceFolder } from "vscode";
 import { TestCase } from "./testCase";
 import { manager } from "./extension";
-import { getDeployTools, getInstance } from "./api/ibmi";
+import { getDeployTools, getInstance } from "./extensions/ibmi";
 import { IBMiTestRunner } from "./runner";
 import { TestingConfig, RUCRTRPG, RUCRTCBL } from "./types";
 import * as path from "path";
